Bump the vcpkg `builtin-baseline`. (#5010)
[SC-48154](https://app.shortcut.com/tiledb-inc/story/48154/2-26-bump-the-vcpkg-builtin-baseline)

This PR bumps the vcpkg `builtin-baseline` which causes all of our
dependencies to be updated to the most recent versions available on
vcpkg. Some of our custom ports are no longer needed and were removed as
well.

To view the full list of updated ports, you can clone vcpkg and run `git
diff 72010900b7cee36cea77aebb97695095c9358eaf
101cc9a69a1061969caf4b73579a34873fdd60fe -- .\versions\baseline.json >
baseline.diff`. This is an abridged summary of the updated ports that we
depend on:

|Port|Old version|New version|
|----|-----------|-----------|
|aws-sdk-cpp|1.11.215|1.11.352|
|azure-core-cpp|1.10.3|1.13.0|
|azure-identity-cpp|1.6.0|1.8.0|
|azure-storage-blobs-cpp|12.10.0|12.11.0|
|azure-storage-common-cpp|12.5.0|12.6.0|
|catch2|3.4.0|3.6.0|
|capnproto|1.0.1|1.0.2|
|curl|8.4.0|8.8.0|
|fmt|10.1.1|10.2.1|
|google-cloud-cpp|2.19.0|2.25.0|
|libwebp|1.3.2|1.4.0|
|libxml2|2.11.6|2.11.7|
|openssl|3.1.4|3.3.1|
|pcre2|10.42|10.43|
|spdlog|1.12.0|1.14.1|
|zlib|1.3|1.3.1|
|zstd|1.5.5|1.5.6|

The bump to the baseline might cause configure errors due to the vcpkg
repository being out of date In this case, try one of the following
things:

* If you are using a cloned vcpkg repository and specified to the Core
either a custom toolchain file, or the `VCPKG_ROOT` environment
variable, update the vcpkg repository to the latest available commit,
and try configuring again.
* If you rely on the Core's build system to automatically download vcpkg
(the default case), try deleting `build/_deps` and configure again. If
it doesn't work, try performing a clean build.

Let me know if any additional testing is required.

---
TYPE: BUILD
DESC: Update vcpkg version baseline to
https://github.com/microsoft/vcpkg/commit/101cc9a69a1061969caf4b73579a34873fdd60fe.