Update GCS emulator.
Should fix CI failures on macOS.
By removing the install python step, the emulator was using the system-installed Python 3.12. However, the emulator had a dependency to an old grpcio version which provided wheels only for Python 3.9, forcing `pip` to build it from source, which failed with compile errors.
The new version depends on a newer `grpcio` which provides wheels for Python 3.8 to 3.12.