Add serialization and API changes for `post_array_schema_from_rest`. (#5237)
This factors out serialization and API changes in #5181 that are
required for the HandleGetArraySchema route. These changes will need to
be available on REST before we can enable the new route for loading the
array schema.

There is a quick summary of the changes required in
[SC-52877](https://app.shortcut.com/tiledb-inc/story/52877/core-serialization-changes-for-loadarrayschema-models).

---
TYPE: IMPROVEMENT
DESC: Add serialization and API changes for post_array_schema_from_rest.