Improve error message of `apply_with_type`. (#5289)
[SC-53336](https://app.shortcut.com/tiledb-inc/story/53336/improve-error-messages-for-unsupported-datatypes-in-aggregate-queries)

If `apply_with_type` (used for aggregates among other things) encounters
an unsupported data type (like non-ASCII strings), it would fail with a
message saying that the type is "not valid". This PR updates it to the
more clear "not supported".

---
TYPE: IMPROVEMENT
DESC: Improve error messages for unsupported data types in aggregates.