Add SC-54473 sum overflow regression test case. (#5283)
<long description>


https://app.shortcut.com/tiledb-inc/story/54473/core-sum-overflow-either-returns-null-or-u64-max-depending-on-tile-capacity

> The tile_metadata has a sum for each tile.
> This function aggregate_tile_with_frag_md combines the sums stored in
the tile metadata.
> The call to update_sum checks for overflow and if it happens sets the
result to be null if it is a nullable result; or sets it to be the type
max if not.
> But if the values inside of the tile themselves have overflowed then
we do not do this at all, the tile metadata has some value (I'm seeing
i64::MAX) and the result is not set to be null.
> I can see a difference in behavior if I take the same set of sparse
data and tweak the schema tile capacity from 1 to 10. With 1 the
overflow is detected but with 10 it is not because the sum is found in
the tile metadata.

This pull request adds a regression test demonstrating that if the tile
capacity is low, overflow is detected; but if the tile capacity is high,
then overflow is not detected because the overflow happened when
computing the tile metadata.

---
TYPE: NO_HISTORY
DESC: Add regression test demonstrating gaps in sum overflow detection.