Dense reader fails early when tile offsets are too large. (#5310)
This change allows to accurately compute the memory usage for tile
offsets before they are loaded. We can then throw an error before we do
any work if the tile offsets are too large for the memory budget. This
will enable us to fail faster on REST when performing large queries and
do fast retries.

Note that the memory usage for tile offsets is done per fragment, which
will ease the implementation of partial tile offsets loading for the
dense reader.

---
TYPE: IMPROVEMENT
DESC: Dense reader fails early when tile offsets are too large.