Pass AWS client configuration to the default credentials provider chain. (#5315)
[SC-55378](https://app.shortcut.com/tiledb-inc/story/55378/ensure-all-aws-clients-and-credentials-providers-are-being-passed-client-configuration)

This PR applies a similar fix to #4641, but for the
`DefaultAWSCredentialsProviderChain` which was being implicitly used
when we don't pass any credentials provider to `S3Client` (it's now
explicitly used).

Because `DefaultAWSCredentialsProviderChain` does not accept a
`ClientConfiguration`, I copied its sources, as well as the sources of
all applicable credentials providers and updated them to accept a
`ClientConfiguration`. This is a temporary measure until the changes to
the providers are upstreamed to the AWS SDK.

You are encouraged to review this PR by looking at each commit
individually.

---
TYPE: BUG
DESC: Fix HTTP requests for AWS default credentials provider chain not
honoring config options.