Add overload to `Array::create` that explicitly accepts a context. (#5325)
[SC-42521](https://app.shortcut.com/tiledb-inc/story/42521/add-overload-to-the-array-create-c-function-that-explicitly-accepts-a-context)

The `Array::create` function uses the context that was used when
creating its `ArraySchema`. This is unintuitive and inflexible, and this
PR fixes it by adding an overload that explicitly accepts a context.

Per discussions the existing overload is not deprecated, but users are
recommended to explicitly pass a context when creating an array whenever
possible.

---
TYPE: CPP_API
DESC: Add overload to the `Array::create` function that explicitly
accepts a context.