Set increased traversal limit in enumeration deserialization (#5365)
So far we'd only apply an increased traversal limit to the
deserialization of certain objects like Query, groups and metadata of
all sorts. For the rest of them, like Array Schema evolution we use the
default Capnp value that is actually
[64MB](https://capnproto.org/encoding.html#:~:text=We%20call%20this%20limit%20the,a%20different%20limit%20if%20desired.).

We had a real life scenario where evolving the array schema hit the
traversal limit and failed. This is most probably due to the addition of
a lot of enumerations in a new schema, and it seems that large
enumerations are commonly used in certain scientific use cases. To cope
with such cases, this PR sets an increased traversal limit to every
deserialization that includes enumerations: schema evolution, array
schema and load enumerations response.

TYPE: BUG
DESC: Set increased traversal limit in enumeration deserialization