Add API to dump groups that returns a string handle. (#5367)
[SC-59017](https://app.shortcut.com/tiledb-inc/story/59017/add-tiledb-group-dump-str-v2-function-that-returns-a-string-handle)

This PR adds a new C API `tiledb_group_dump_str_v2` that dumps a group
into a `tiledb_string_t*`. This allows the string to be safely freed
under certain circumstances (such as when debug and release CRT are
mixed on Windows).

The existing `tiledb_group_dump_str` API was deprecated. Migrating to
the new API is straightforward and requires no changes to the public
surface of the higher-level APIs. The C++ API and the C groups example
are migrated in this PR.

---
TYPE: C_API
DESC: Added `tiledb_group_dump_str_v2` API that returns a
`tiledb_string_t*`. The existing `tiledb_group_dump_str` API is
deprecated.