Bug fix: consolidation failure on dense arrays without respecting current domain (#5390)
This PR addresses a bug in the `Domain::expand_to_tiles(NDRange*)`
function, which previously did not respect the current domain. The lack
of this consideration led to errors when the tile extent was larger than
the current domain, causing out-of-bounds issues during consolidation.
To resolve this, the PR introduces a new function that appropriately
accounts for the current domain.

[sc-59793] and
https://github.com/single-cell-data/TileDB-SOMA/issues/3383

Huge kudos to @-johnkerl for getting 99.9% of this done.

---
TYPE: BUG
DESC: This PR fixes a bug in `Domain::expand_to_tiles(NDRange*)`, which
failed to respect the current domain, causing errors when tile extents
exceeded it. A new function is introduced to properly account for the
current domain during consolidation.

---------

Co-authored-by: John Kerl <kerl.john.r@gmail.com>