Add endpoint for REST version and capabilities. (#5429)
This adds the REST capabilities endpoint to `RestClientRemote` and
serialization for the `RestCapabilities` response. The endpoint is
currently not in use internally and there is no way to make the request
using public APIs.

The `RestCapabilities` response currently provides the TileDB core
version deployed to REST, and the minimum TileDB core client version
supported by the REST server. The response may be updated in the future
to add more information as needed - [discussion in
shortcut](https://app.shortcut.com/tiledb-inc/story/61403/rest-route-to-check-server-side-libtiledb-version-and-capabilities#activity-62215)
and [design
document](https://www.notion.so/Design-REST-route-to-check-server-side-libtiledb-version-181326501175800ea184eb6909b65d20?pvs=4)

---
TYPE: FEATURE
DESC: Add endpoint for REST version and capabilities.

---------

Co-authored-by: Theodore Tsirpanis <teo@tsirpanis.gr>