Do not use read-ahead cache in array/group directory. (#5452)
[SC-62900](https://app.shortcut.com/tiledb-inc/story/62900)

In all cases where we read files from the VFS there, we get the file's
size beforehand and read the entirety of it. This makes the read-ahead
cache not actually required.

Fixes failures when reading empty `.vac` files (in S3 we do a
`GetObject` with a range of `0-102399`, which fails with a range not
satisfiable error).

---
TYPE: BUG
DESC: Fix failures when opening an array or group with an empty `.vac`
file.