Do not create empty .vac files in metadata consolidation (#5453)
When we are trying to consolidate array or group metadata but there are
no array or group metadata present, we are creating an empty `.vac` file
on local filesystems. In object stores, notably on S3 that I have
tested, our `write` method currently doesn't create a file when size is
`0` (this is most probably another issue to fix), so we don't observe
the same behavior.

This PR unifies the behavior by enforcing that we don't create a `.vac`
file if no consolidation is possible, aka if there are 0 or 1
array/group metadata present. We already do such a check for fragment
metadata, fragments and commit consolidation.
Both tests in this PR were failing without the fix and are now passing.

---
TYPE: BUG
DESC: Do not create empty .vac files in metadata consolidation

(cherry picked from commit a22fe138f4785180e8c5b8f3c742bca4cae89496)