chore: [sc-63841] [core] adding multiple attributes with schema evolution does not add them in a predictable order (#5470)
Story details: https://app.shortcut.com/tiledb-inc/story/63841

Previously attributes were stored in the `SchemaEvolution` via
`std::unordered_map` which resulted in an arbitrary order that the
attributes would be appended to the target schema.

This pull request uses `std::vector` instead to preserve the order of
added attributes. Lookups by attribute name now require a linear scan of
the added attributes. This will probably be faster than the
`unordered_map` lookup was for small numbers of attributes; when there
are many attributes it may not be, but it should be a rare use case
where that would be noticeable.

I also noticed that it is possible to evolve into an invalid schema
which cannot be loaded. This pull request adds an integrity check before
returning a successful schema evolution to fix this.

---
TYPE: IMPROVEMENT
DESC: Schema evolution consistent attribute order