Optimize `VFS::touch` and fix defects and race conditions. (#5285)
[SC-53188](https://app.shortcut.com/tiledb-inc/story/53188/azure-update-touch-to-use-conditional-requests)

[SC-53189](https://app.shortcut.com/tiledb-inc/story/53189/gcs-update-touch-to-use-conditional-requests)

[SC-54494](https://app.shortcut.com/tiledb-inc/story/54494/win-touch-is-prone-to-race-conditions)

This PR updates the implementations of `VFS::touch` for Azure, GCS and
Windows to improve performance and fix race conditions:

* On Azure we remove checking if the blob exists, and instead use a
single [conditional
request](https://learn.microsoft.com/en-us/rest/api/storageservices/specifying-conditional-headers-for-blob-service-operations)
to upload an empty blob if it does not already exist. This reduces
round-trips and fixes race conditions if the blob was created between
the existence check and the upload.
* GCS did not even have an existence check, which means that the
behavior of `touch` was incorrect and always overwrote the file. This
was fixed by adding a [request
precondition](https://cloud.google.com/storage/docs/request-preconditions).
* On Windows the existence check was removed. We were already doing the
right thing and opened the file with `CREATE_NEW`, but did not mask
failures with `ERROR_FILE_EXISTS` status code. We fix this.

Additionally, the documentation of `tiledb_vfs_touch` was updated to
mention that the file is created only if it does not exist, matching
existing behavior in most VFS implmentations, and the expectations from
the `touch` command.

Testing coverage was added.

---
TYPE: BUG
DESC: Fixed `tiledb_vfs_touch` to not overwrite existing files on GCS.

---
TYPE: BUG
DESC: Fixed `tiledb_vfs_touch` to not overwrite existing files or fail
on Azure and Windows respectively, under race conditions.

---
TYPE: C_API
DESC: Update documentation of `tiledb_vfs_touch` to specify that
existing files are not overwritten, matching the current behavior.