Fix opening arrays with unrecognized files in their root directory et al. (#5474)
[SC-64121](https://app.shortcut.com/tiledb-inc/story/64121/cannot-open-array-with-unrecognized-files-in-its-root-directory)

This PR updates `ArrayDirectory::is_fragment` to return a negative
result if the file's name is not a fragment identifier, instead of
throwing and failing the open.

I also did a small clean-up of the `FragmentID` class.

---
TYPE: BUG
DESC: Fixed opening arrays with unrecognized files in their root
directory.