Add test for dropping all attributes which reference an enumeration (#5478)
I observed this behavior and asked about it in slack and the consensus
was that we should allow dangling enumerations. This pull request
updates a comment for an existing "dangling enumeration" test and adds
another test which leaves behind a dangling enumeration.

---
TYPE: NO_HISTORY
DESC: Add test for dropping all attributes which reference an
enumeration