Prevent undefined behavior in Enumeration constructor  (#5479)
Constructing a string enumeration with only an empty string as a value
can result in UB inside the enumeration constructor. There
is an acceptable parameter configuration (`data == nullptr` and
`data_size == 0`) where both the internal `Buffer`'s pointer and the
data pointer are null. Writing to the buffer results in
`std::memcpy(nullptr, nullptr, 0)` which is UB according to the spec.

The same holds true for the offsets buffer in fixed sized enumerations.

The data size check in `generate_value_map` is now redundant because the
data buffer when the enumeration in not empty will have a valid memory
address thus the `string_view` construction will not exhibit UB when
constructed using an empty string. IN case the enumeration is empty the
loop limits will guard from dereferencing the null buffers

---
TYPE: BUG
DESC: Fix undefined behavior in Enumeration constructor