Add `Enumeration::index_of` API (#5482)
Add `Enumeration::index_of` API which returns an
`std::optional<uint64_t>` containing the internal index of a value in an
Enumeration if it exists.

User code who needs to extend an enumeration given a list of values need
to recreate the internal hash map structure to efficiently find which
values are missing from the enumeration. Also remapping the indexes from
a subset of values to the values the enumeration holds also requires the
creation of the values hash map. Providing access to the internal hash
map of the enumeration eliminates the need to recreate the map which
might me an expensive step.

---
TYPE: FEATURE 
DESC: Add `Enumeration::index_of` API