Fix SparseGlobalOrderReader check for progression (#5485)
Resolves
[SC-64765](https://app.shortcut.com/tiledb-inc/story/64765/400-s-in-rapidcheck-rest-ci-tests).

The linked story observed many 400 errors occurring during the REST
server rapidcheck CI job. This prompted two questions: why were these
errors happening? And why did they not cause the tests to fail?

The errors were happening because the rapidcheck query condition
generator had a mistake in it, causing it two insert two values rather
than one into its AST nodes; and then a bug in the rapidcheck property
was erroneously swallowing the errors, causing them not to be reported
to the test handler.

These two issues masked an error in the sparse global order reader where
it would throw an exception of inability to progress if all of the
loaded tiles did not pass a query condition.

This pull request fixes all of the above issues and also makes some
improvements to related test support code.

Each commit is a self-contained item so it might be helpful to review
this one commit at a time.

---
TYPE: BUG
DESC: Fix sparse global order reader progress check

---------

Co-authored-by: Ryan Roelke <ryan@tiledb.com>