Support setting of schema-creation timestamp (#5484)
This work builds on the work originally done by @johnkerl in
https://github.com/TileDB-Inc/TileDB/pull/5260 which introduced the
`tiledb_array_schema_alloc_at_timestamp` to the C API. As iterated from
the previous PR, this removes the CPP API addition and adds unit testing
for creating a schema at specific timestamps and further schema
evolution at specific timestamps for adding and dropping attributes,
extending enumerations, and resizing the current domain.

The motivation for the PR is outlined in more detail
[here](https://app.shortcut.com/tiledb-inc/story/47660/core-c-c-api-support-to-set-initial-array-schema-timestamp#activity-53864).
TileDB-SOMA has had a longstanding
[support](https://github.com/single-cell-data/TileDB-SOMA/pull/1001) to
create `SOMAArray`s at a given `tiledb_timestamp`. However, there is no
way to set the initial `ArraySchema` file timestamp, and it is always
written to the current timestamp. This limitation causes issues for
performing schema evolutions at a timestamp, specifically with
[extending
enumerations](https://github.com/single-cell-data/TileDB-SOMA/issues/2879)
and [resizing
arrays](https://app.shortcut.com/tiledb-inc/story/61254/python-array-resize-and-time-travel-not-working-correctly).

To discourage misuse of this feature, we have limited its availability
to the C API and added a warning to emphasize that is to be used with
caution.

---
TYPE: FEATURE | C_API 
DESC: Support setting of schema-creation timestamp

---------

Co-authored-by: John Kerl <kerl.john.r@gmail.com>
Co-authored-by: Theodore Tsirpanis <teo@tsirpanis.gr>
Co-authored-by: Beka Davis <31743465+bekadavis9@users.noreply.github.com>
Co-authored-by: Ypatia Tsavliri <ypatia@tiledb.com>
Co-authored-by: Isaiah Norton <isaiah@tiledb.io>