Fix TILEDB_ASSERTIONS build configuration, enabling CI jobs to run in both debug and release builds with assertions on (#5514)
While working on #5510 I discovered the following:
- the standalone unit tests are only built in debug mode, with
`TILEDB_ASSERTIONS=ON`.
- `tiledb_unit` is only built in release mode, with
`TILEDB_ASSERTIONS=OFF`.
- the standalone unit tests _cannot_ be built in release mode with
`TILEDB_ASSERTIONS=ON`.
- `tiledb_unit` has at least one deterministic assertion failure when
built in debug mode.

This pull request intends to address these problems. There are three
configurations which we should be testing, but are not properly doing:
1) debug builds with assertions on, i.e. the development process.
2) release builds with assertions on, i.e. the continuous integration
make-sure-we-didn't-break-anything process.
3) release builds with assertions off, i.e the maximum performance
package we want to deliver to customers.

The intent of `TILEDB_ASSERTIONS` is to enable our build to distinguish
between (2) and (3). As we see above, we are failing this intent, and
failing to validate the three common build configurations.

The goal of this pull request is to fix `TILEDB_ASSERTIONS` so that each
of these build configurations is tested at some point in our CI
pipelines. We believe it is sufficient to test only (2) as part of pull
request CI; whereas (1) and (3) can be tested in nightlies.

To fix `TILEDB_ASSERTIONS`, we will avoid using `assert` entirely,
preferring custom assertion routines. See `tiledb/common/assert.h`
introduced in this PR for an elaboration as to why this is what we want
to do.

This pull request introduces those assertion routines `iassert` and
`passert`; updates all of the TileDB internal code to use them instead
of vanilla `assert`; adds a lint to ensure that we do not use vanilla
`assert` in the future; and then updates CI to test configuration (2)
for pull requests, and (1), (2) and (3) for nightlies.

---
TYPE: NO_HISTORY
DESC: Fix `TILEDB_ASSERTIONS` build configuration and update CI
accordingly

---------

Co-authored-by: Theodore Tsirpanis <teo@tsirpanis.gr>