Bugfix: Erase `param` from `set_params_` in `Config::unset`. (#5519)
This PR updates the `Config::unset` method to ensure consistency between
`param_values_` and `set_params_`. Previously, when a parameter was
unset, it was removed or reset in `param_values_` but not in
`set_params_`. This could lead to stale entries in `set_params_`.

---
TYPE: BUG
DESC: Erase `param` from `set_params_` in `Config::unset`.