Fix standalone unit test build with `-DTILEDB_ASSERTIONS=ON` (#5521)
Immediately after merging #5514 I noticed that despite the claims made
in the PR description the changes failed to add `-DTILEDB_ASSERTIONS=ON`
to the standalone unit tests. This became apparent because they failed
to build!

This pull request fixes that. We split some definitions out into
`assert.c` for an `assert` object library which propagates adding the
`fmt::fmt` library to the link and include paths.

---
TYPE: NO_HISTORY
DESC: Fix standalone unit test build with `-DTILEDB_ASSERTIONS=ON`

---------

Co-authored-by: Ypatia Tsavliri <ypatia@tiledb.com>