Add support for Profiles onto the Config. (#5498)
This PR adds support for `Profiles` onto the `Config` by introducing the
corresponding `set_profile` method to the internal class, as well as to
both the C and C++ APIs. `Profiles` created with the default `name` and
`dir` are automatically loaded, without the need to call the
`set_profile` method. The `Profiles` serve as a third fallback mechanism
in the `Config` when retrieving parameters - after user-set config
values and environment variables.

---

[sc-64254]

---
TYPE: FEATURE
DESC: Add support for `Profiles` onto the `Config`.

---------

Co-authored-by: Agisilaos Kounelis <kounelisagis@gmail.com>