Fix windows runtime out of bounds assert for empty output tiles in sparse global order reader (#5523)
This pull request corrects a `std::vector` buffer access which fails an
assertion inside the windows C runtime.
```
    future.merge_ = algorithm::parallel_merge(
        resources_.compute_tp(),
        merge_resources,
        future.merge_options_,
        future.fragment_result_tiles_,
        *static_cast<ResultTileCmp*>(future.cmp_.get()),
        &preprocess_tile_order_.tiles_[0]);
```
The `&preprocess_tile_order_.tiles_[0]` access intends to get a pointer
to a buffer for the parallel merge to write its output. When the
expected number of results is zero, however, this represents is an
out-of-bounds access, even if we would never write to that memory
location.

Our resolution to this issue draws inspiration from this safety feature.
We use `std::span` in the parallel merge instead of a raw pointer, and
assert that the writes are in bounds.

---
TYPE: NO_HISTORY
DESC: Fix windows runtime out of bounds assert for empty output tiles in
sparse global order reader