Fix segfault race condition at process exit (#5477)
I discovered a couple race conditions leading to segfaults in a separate
repository. After debugging I managed to narrow this down to issues
around the lifetimes of our static global state instances. As these
instances are all destructed as the main thread exits, anything in
separate threads can end up attempting to reuse these instances after
they have been destructed.

The test included in this PR doesn't actually use threads due to
flakiness concerns. Instead we can observe the same phenomenon by
attempting to store a `static std::optional<Context>` instance which
ends up having its destructor run after the destructors for GlobalState
and the root Logger.

This change updates the GlobalState to be a `static
std::shared_ptr<GlobalState>` and the `Logger` to be a `static Logger*`.
In the case of `GlobalState` we then just rely on shared_ptr reference
counting to ensure that the instance is alive for as long as a `Context`
needs it. The `Logger*` on the other hand is just never delete'd so that
any log messages generated are included rather than silencing any thread
after the main thread exits.

---
TYPE: BUG
DESC: Fix segfault race condition at process exit

---------

Co-authored-by: Ryan Roelke <ryan@tiledb.com>