chore: add callback registration to `passert_failure` to enable dumping rapidcheck test input (#5532)
Resolves CORE-233.

A `passert` failure crashes the process without returning. While
`Catch2` has a `SIGABRT` handler to print some details around this,
`rapidcheck` does not. As such, if a randomized `rapidcheck` input fails
a `passert`, no details about what the input was are printed.

Ostensibly we have some means of recovering that failed input using the
seed which produced the test case, but this might not be helpful:
1) we don't know if it is platform-independent
2) if there is a non-deterministic or undefined behavior leading to
failure, then the seed may not be enough to clue us in

This pull request adds a callback registration to `passert_failure` so
that we can register a callback in our `rapidcheck` properties to print
the last attempted input in the event of a `passert` failure. This way
we can recover a bad input with certainty.

Example:
```
FATAL TileDB core library internal error: false
  test/src/unit-sparse-global-order-reader.cc:1396
Last rapidcheck input: (2, 1, 1, [[1, 1]], d1 LT 0x01 0x00 0x00 0x00)
```

---
TYPE: NO_HISTORY
DESC: add callback registration to `passert_failure` to enable dumping
rapidcheck test input