chore: add memory tracker `passert_failure` callback to dump memory info to `stderr` if destructor assertion fails (#5535)
We are occasionally seeing the following failure in CI:
```
106: FATAL TileDB core library internal error: total_counter_.fetch_add(0) == 0 && "MemoryTracker destructed with outstanding allocations."
106:   tiledb/common/memory_tracker.cc:192
```
This assert failing indicates that we have some memory resource which is
allocating memory but not freeing it. However, we receive no guidance as
to what entity might be responsible for that.

This pull request adds a `passert_failure` callback which dumps the
memory tracker state to `stderr` if this assertion fails. This way we
will have a little more information the next time it fails.

Here's an example output from this assertion failing in a rapidcheck
test:
```
FATAL TileDB core library internal error: total_counter_.fetch_add(0) == 0 && "MemoryTracker destructed with outstanding allocations."
  tiledb/common/memory_tracker.cc:192
LAST RAPIDCHECK INPUT:
(200, 8, 2, [], ???
)
MEMORY REPORT:
{"by_type":{"FilteredData":0,"FilteredDataBlock":0,"ParallelMergeControl":0,"QueryCondition":0,"ResultTileBitmap":0,"TileData":0,"TileHilbertValues":0},"total_memory":1,"tracker_id":"23","tracker_type":"QueryRead"}
```

---
TYPE: NO_HISTORY
DESC: Add callback to print memory tracker info if its destructor assert
fails