chore: update build and CI to integrate Rust components (#5528)
This pull request kicks off building TileDB core with Rust components.
We add a simple Rust crate as an object library `arithmetic_rs` and use
it so far only to test existing C++ code. The intent is to incrementally
merge the Rust build and CI work separately from the non-trivial changes
we expect to use it for.

We add a new build configuration option `TILEDB_RUST` and only build and
link the Rust components if it is off (which is the default).

At the moment the Rust code is only used in the standalone unit tests.

As for the build process... we add a function `add_crate` which
registers a Rust crate as an object library. This function accepts a
list of Rust source files which have a `#[cxx::bridge]`-annotated module
- the build process generates C++ header files for these modules. The
generated files are placed at a rather ugly path inside of
`${CMAKE_BINARY_DIR}/tiledb/oxidize/target`, so a non-trivial amount of
scaffolding is written in order to sanitize the include paths - for
example, the `unit_arithmetic.cc` file in this example can do `#include
"tiledb/oxidize/arithmetic.h"` instead of `#include
"build/tiledb/oxidize/target/cxxbridge/expr/lib.rs.h"` or something like
that.

---
TYPE: NO_HISTORY
DESC: Update build and CI to handle Rust components